requirements.txt
# Updated requirements.txt compatible with Python 3.12
# Head Pose Estimation Project

# Core dependencies
opencv-python>=4.8.0
numpy>=1.26.0  # Compatible with Python 3.12
torch>=2.0.0
torchvision>=0.15.0

# MediaPipe (fixed versions)
mediapipe==0.10.21
protobuf==4.25.3

# Data processing and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pandas>=2.0.0
tqdm>=4.66.0

# Configuration
pyyaml>=6.0.0

# Optional: TensorBoard for logging
tensorboard>=2.13.0

# Testing (optional)
pytest>=7.4.0